Add info message on system settings save
A new variable screenMsg is added to the Application class.
Whenever that var contains a non-empty string, that string
will be displayed on the bottom left corner of the window
during 2 seconds.

At the moment it is used to inform about system settings save,
but it can be used for any other message just by assigning any
string to screenMsg var. The message will be displayed for two
seconds on the bottom left corner.