Move systems.json parsing to Cache
File is now loaded in memory and should be returned from memory
onwards.
Still not unified in terms of method naming.
The method to update the selected exex still does it directly
in the file instead of in memory (so systems cache won't reflect
the change). Read the FIXME at the beggining of the method.